Moved all ZK-related tests to its own project in main solution
