+++ test/Tester/Tester.csproj
@@ -185,10 +185,6 @@
       <Project>{40ee3b00-d381-485f-9c69-ff706837deed}</Project>
       <Name>OrleansTestingHost</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\src\OrleansZooKeeperUtils\OrleansZooKeeperUtils.csproj">
-      <Project>{45918dd2-53d9-4b27-bc6b-17c197dbc645}</Project>
-      <Name>OrleansZooKeeperUtils</Name>
-    </ProjectReference>
     <ProjectReference Include="..\..\src\Orleans\Orleans.csproj">
       <Project>{bc1bd60c-e7d8-4452-a21c-290aec8e2e74}</Project>
       <Name>Orleans</Name>
