+++ test/TesterInternal/TesterInternal.csproj
@@ -116,7 +116,6 @@
     <Compile Include="StorageTests\StorageProviders\BaseJSONStorageProvider.cs" />
     <Compile Include="StorageTests\StorageProviders\FileStorageProvider.cs" />
     <Compile Include="StorageTests\StorageProviders\IJSONStateDataManager.cs" />
-    <Compile Include="MembershipTests\ZookeeperMembershipTableTests.cs" />
     <Compile Include="StreamingTests\DynamicStreamProviderConfigurationTests.cs" />
     <Compile Include="StreamingTests\HaloStreamSubscribeTests.cs" />
     <Compile Include="StreamingTests\MultipleStreamsTestRunner.cs" />
@@ -172,10 +171,6 @@
     <ProjectReference Include="..\..\src\OrleansTestingHost\OrleansTestingHost.csproj">
       <Project>{40ee3b00-d381-485f-9c69-ff706837deed}</Project>
       <Name>OrleansTestingHost</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\src\OrleansZooKeeperUtils\OrleansZooKeeperUtils.csproj">
-      <Project>{45918dd2-53d9-4b27-bc6b-17c197dbc645}</Project>
-      <Name>OrleansZooKeeperUtils</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\src\Orleans\Orleans.csproj">
       <Project>{bc1bd60c-e7d8-4452-a21c-290aec8e2e74}</Project>
