+++ test/TesterZooKeeperUtils/project.json
@@ -1,0 +1,16 @@
+﻿{
+  "dependencies": {
+    "FluentAssertions": "4.0.0",
+    "Newtonsoft.Json": "9.0.1",
+    "xunit": "2.1.0",
+    "xunit.runner.visualstudio": "2.1.0",
+    "Xunit.SkippableFact": "1.2.14"
+  },
+  "frameworks": {
+    "net451": {}
+  },
+  "runtimes": {
+    "win-anycpu": {},
+    "win": {}
+  }
+}
